pyproject.toml
[build-system]
requires = ["setuptools >= 67.7", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "python-can-coe"
description = "An interface introducing CAN over Ethernet into python-can"
dynamic = ["version"]
dependencies = [
    "python-can>=4.5.0",
]
requires-python = ">=3.8"
authors = [
  {name = "Gerrit Beine", email = "mail@gerritbeine.com"},
]
maintainers = [
  {name = "Gerrit Beine", email = "mail@gerritbeine.com"},
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.8",
]

[project.urls]
Homepage = "https://c0d3.sh/smarthome/python-can-coe"

[project.entry-points."can.interface"]
coe = "coe.coe:CoE"

[tool.setuptools.packages.find]
include = ["coe*"]

